pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.45",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.5",
    "pydantic[email]>=2.12.5",
    "uvicorn[standard]>=0.32.1",
    "docling>=2.64.1",
    "python-multipart>=0.0.20",
    "langchain-groq>=1.1.1",
    "groq>=0.37.1",
    "black>=25.12.0",
    "isort>=7.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
